main.js стал модулем в html и старый main разбит на модули(util, data)
